charm/pyproject.toml
[project]
name = "release-monitor-gcloud-charm"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "ops>=3,<4",
    "PyYAML>=6.0",
]

[dependency-groups]
lint = [
    "ruff",
    "codespell",
    "pyright",
]
unit = [
    "coverage[toml]",
    "ops[testing]",
    "pytest",
]
integration = [
    "juju",
    "pytest",
    "pytest-operator",
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]

[tool.coverage.run]
branch = true

[tool.coverage.report]
show_missing = true
